requirements.txt
# Core dependencies
fastapi
uvicorn
pydantic
pymupdf
python-docx
faiss-cpu
python-multipart
python-dotenv

# Local inference dependencies (with Qwen3-Embedding support)
torch
transformers>=4.51.0
sentence-transformers>=2.7.0
accelerate
bitsandbytes
psutil
numpy
